bus: refactor client
